Refactors CameraScreen using a textDetector service
