+++ services/textDetector.js
@@ -4,7 +4,42 @@ class TextDetector {
   constructor(photo) {
     this.photo = photo
   }
-  
+
+  postOptions = (body) => {
+    return {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify(body)
+    }
+  }
+
+  textDetection = () => {
+    let body = {
+      "requests":[
+        {
+          "image":{
+            "source":{
+              "imageUri": `https://s3.us-east-2.amazonaws.com/my-hep-images/photos/${this.photo}`
+            }
+          },
+          "features":[
+            {
+              "type":"TEXT_DETECTION",
+              "maxResults":1
+            }
+          ]
+        }
+      ]
+    }
+    fetch(`https://vision.googleapis.com/v1/images:annotate?key=${config['GOOGLE_VISION_KEY']}`,
+      this.postOptions(body))
+      .then(result => result.json())
+      .then(response => {
+        let text = response.responses[0].textAnnotations[0].description
+        console.warn(text)
+      })
+      .catch(error => console.warn(error))
+  }
 }
 
 module.exports = TextDetector
